fix(ct): CT.23 structure budget, deadcode, and E2E coverage gate
Merge probe action handlers to keep contenttools under the package-size
budget, remove unused flashcards helpers, and register CT.23 in the
completed-feature coverage manifest.
