+++ server/internal/service/contenttools/tools/flashcards/state.go
@@ -167,66 +167,3 @@ func EndActiveSession(st *State, now string) SessionRecord {
 	st.ActiveSession = nil
 	return rec
 }
-
-// MergeStates merges concurrent flashcard states (max seen, last-write ratings).
-func MergeStates(a, b State) State {
-	out := EmptyState()
-	if a.V != 0 {
-		out.V = a.V
-	}
-	if b.V != 0 {
-		out.V = b.V
-	}
-	out.Cards = map[string]CardProgress{}
-	keys := map[string]struct{}{}
-	for k := range a.Cards {
-		keys[k] = struct{}{}
-	}
-	for k := range b.Cards {
-		keys[k] = struct{}{}
-	}
-	for k := range keys {
-		pa, okA := a.Cards[k]
-		pb, okB := b.Cards[k]
-		switch {
-		case okA && !okB:
-			out.Cards[k] = pa
-		case okB && !okA:
-			out.Cards[k] = pb
-		default:
-			merged := pa
-			if pb.Seen > merged.Seen {
-				merged.Seen = pb.Seen
-			}
-			if pb.LastSeenAt > merged.LastSeenAt {
-				merged.LastSeenAt = pb.LastSeenAt
-				merged.LastRating = pb.LastRating
-			}
-			if merged.FirstRating == nil {
-				merged.FirstRating = pb.FirstRating
-			}
-			out.Cards[k] = merged
-		}
-	}
-	// Prefer longer session history.
-	if len(b.Sessions) > len(a.Sessions) {
-		out.Sessions = append([]SessionRecord{}, b.Sessions...)
-	} else {
-		out.Sessions = append([]SessionRecord{}, a.Sessions...)
-	}
-	if a.FirstPassCompletedAt != "" {
-		out.FirstPassCompletedAt = a.FirstPassCompletedAt
-	}
-	if b.FirstPassCompletedAt != "" && (out.FirstPassCompletedAt == "" || b.FirstPassCompletedAt < out.FirstPassCompletedAt) {
-		out.FirstPassCompletedAt = b.FirstPassCompletedAt
-	}
-	// Keep the more advanced active session if any.
-	if a.ActiveSession != nil && (b.ActiveSession == nil || a.ActiveSession.Reviewed >= b.ActiveSession.Reviewed) {
-		cp := *a.ActiveSession
-		out.ActiveSession = &cp
-	} else if b.ActiveSession != nil {
-		cp := *b.ActiveSession
-		out.ActiveSession = &cp
-	}
-	return out
-}
